C0918242230
- continue coding for exercise 154
