+++ main.py
@@ -1,23 +1,32 @@
-from faker import Faker
-import random
-
-def generate_names_and_numbers(gender, year):
-    fake = Faker()
-    names = set()
-    while len(names) < 100:
-        if gender == 'female':
-            names.add(fake.first_name_female())
-        else:
-            names.add(fake.first_name_male())
-
-    with open(f'{year}{gender[0]}', 'w') as file:
-        for name in names:
-            number = random.randint(1, 99)
-            file.write(f'{name},{number}\n')
-
-years = ['2018', '2019', '2020']
-genders = ['female', 'male']
-
-for year in years:
-    for gender in genders:
-        generate_names_and_numbers(gender, year)
+#!/bin/bash
+#*************************************************************
+# Date: 070324 (Expected Solution with 50 Lines of Code)     *
+# Title: Names that Reached Number One                       *
+# Status: In Progress (In Progress / Testing / Working)      *
+#  The baby names data set consists of over 200 files. Each  *
+# file contains a list of 100 names, along with the number of*
+# times each name was used. There are two files for each     *
+# year: one containing names used for girls and the other    *
+# containing names used for boys. The data set includes data *
+# for every year from 1900 to 2012. Write a program that     *
+# reads every file in the data set and identifies all of the *
+# names that were most popular in at least one year. Your    *
+# program should output two lists: one containing the most   *
+# popular names for boys and the other containing the most   *
+# popular names for girls. Neither of your lists should      *
+# include any repeated values.                               *
+#                                                            *
+# Computed Result Validated:                                 *
+#*************************************************************
+#-------------------------------------------------------------
+def func_name(user_in):
+  return
+#-------------------------------------------------------------
+if __name__ == "__main__":
+  file_location = input("Enter file location: ")
+  file_name = input("Enter file name: ")
+  file_path = f"{file_location}/{file_name}"
+  file_handle = open(file_path, "r")
+  file_data = file_handle.read()
+  print("Thank you for using this app.")
+#*************************************************************
